rename spans to segment in utils
